requirements.txt
coreapi==2.1.1
coverage==4.3.1
Django==1.8
django-filter==0.15.3
djangorestframework==3.5.3
factory-boy==2.7.0
fake-factory==0.7.2
geojson==1.3.3
ipaddress==1.0.17
itypes==1.1.0
Jinja2==2.8.1
Markdown==2.6.7
MarkupSafe==0.23
psycopg2==2.6.2
python-dateutil==2.6.0
requests==2.12.4
six==1.10.0
uritemplate==3.0.0
